@@ -1,6 +1,6 @@
 import { CdkTextareaAutosize } from '@angular/cdk/text-field';
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { FormGroup, FormBuilder, FormControl, Validators } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormBuilder, UntypedFormControl, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ThinkYe } from 'src/app/models/classes/think-ye';
 import { FirestoreService } from 'src/app/services/firestore.service';
@@ -14,13 +14,13 @@ export class YourThoughtsComponent implements OnInit {
   
   @ViewChild('autosize') autosize: CdkTextareaAutosize;
   
-  inputForm: FormGroup;
+  inputForm: UntypedFormGroup;
   buttonDisabled: boolean;
   loading: boolean;
   submitText: string;
   thinkYe: ThinkYe;
 
-  constructor(private fb: FormBuilder, private route: ActivatedRoute, private firestoreService: FirestoreService, private router: Router) {
+  constructor(private fb: UntypedFormBuilder, private route: ActivatedRoute, private firestoreService: FirestoreService, private router: Router) {
     const thinkYeId = route.snapshot.url[0].path;
     firestoreService.getThinkYe(thinkYeId).subscribe(result => this.thinkYe = result);
   }
@@ -28,7 +28,7 @@ export class YourThoughtsComponent implements OnInit {
   ngOnInit(): void {
     this.setLoadingState(false);
     this.inputForm = this.fb.group({
-      thought: new FormControl('', [Validators.required]),
+      thought: new UntypedFormControl('', [Validators.required]),
     });
   }
 
